refactored itemprop to take in an item and display it without any fetches
