+++ .env
@@ -9,4 +9,4 @@ cloudinary.config({
   api_secret = 'LVNgD7VCheNOkp5714CvuCVAMWI' 
 
 # REACT_APP_SERVER_URL = "https://traderz-post.herokuapp.com"
-REACT_APP_SERVER_URL = "http://localhost:3003"
+REACT_APP_SERVER_URL = "http://localhost:3001"
